requirements.txt
Pillow>=9.0
beautifulsoup4>=4.9
